backend/requirements.txt
fastapi==0.115.6
python-multipart==0.0.17
uvicorn[standard]==0.32.1
gunicorn==23.0.0
python-dotenv==1.0.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.1
openai==1.59.7
pypdf==5.1.0
python-docx==1.1.2
